+++ src/dipdup/config.py
@@ -872,11 +872,6 @@ class HasuraConfig:
             raise ConfigurationError(f'`{v}` is not a valid Hasura URL')
         return v.rstrip('/')
 
-    @validator('source', allow_reuse=True)
-    def valid_source(cls, v):
-        if v != 'default':
-            raise NotImplementedError('Multiple Hasura sources are not supported at the moment')
-
     @cached_property
     def headers(self) -> Dict[str, str]:
         if self.admin_secret:
@@ -1043,7 +1038,6 @@ class DipDupConfig:
         paths: List[str],
         environment: bool = True,
     ) -> 'DipDupConfig':
-
         current_workdir = os.path.join(os.getcwd())
 
         json_config: Dict[str, Any] = {}
@@ -1052,8 +1046,11 @@ class DipDupConfig:
             path = os.path.join(current_workdir, path)
 
             _logger.debug('Loading config from %s', path)
-            with open(path) as file:
-                raw_config = file.read()
+            try:
+                with open(path) as file:
+                    raw_config = file.read()
+            except OSError as e:
+                raise ConfigurationError(str(e))
 
             if environment:
                 _logger.debug('Substituting environment variables')